This will send the humidity and temperature to TTN
With this change you directly send the data to thethingsnetwork.org. Please note the Payload Formatter in the description which needs to be added into the TTN console, to change the sent data from bytes to numbers. I had to read out the sensor again and stored the data in new variables, I'm fairly sure this can be solved nicer, but this works too.